gpu fixes
Former-commit-id: 048b415adfb8ed52d82631e3f4230a2f7fea85c6

+++ src/gpu/include/cml/cml_defs.cuh
@@ -5,9 +5,9 @@ namespace cml {
 
 typedef unsigned int uint;
 
-const uint kTileSize = 32u;
-const uint kBlockSize = 256u;
-const uint kMaxGridSize = 65535u;
+const size_t kTileSize = 32u;
+const size_t kBlockSize = 256u;
+const size_t kMaxGridSize = 65535u;
 
 }  // namespace cml
